server/requirements.txt
flask==1.0.2
flask-cors==3.0.7
mysql-connector
